// src/utils/types.ts
export interface Settings {
  buttonIcon: string;
  headerBgColor: string;
  chatBotBorderRadius: string;
  borderColor: string;
  botTextColor: string;
  userChatTextColor: string;
  botBubbleBgColor: string;
  userBubbleBgColor: string;
  textFont: string;
}

export type fontOptionsTypes = {
  name: string;
  fontFamily: string;
};